Added endpoint to get data of one vessel

+++ src/main/java/pl/devcezz/barentswatch/api/OpenPosition.java
@@ -2,7 +2,7 @@ package pl.devcezz.barentswatch.api;
 
 import java.util.List;
 
-public record OpenPosition(String timeStamp, Long mmsi, Geometry geometry) {
+public record OpenPosition(String timeStamp, Integer mmsi, Geometry geometry) {
 
     record Geometry(String type, List<Double> coordinates) {
     }
